Extract toggle mute state functions to own hooks
