+++ packages/react-sdk/index.ts
@@ -11,6 +11,8 @@ export * from './src/types';
 export {
   useHorizontalScrollPosition,
   useVerticalScrollPosition,
+  useToggleAudioMuteState,
+  useToggleVideoMuteState,
 } from './src/hooks';
 
 // TODO: set valid version
