+++ packages/react-sdk/src/components/CallControls/ToggleAudioButton.tsx
@@ -1,9 +1,7 @@
-import { ComponentType, useCallback, useEffect, useState } from 'react';
+import { ComponentType } from 'react';
 import { OwnCapability, SfuModels } from '@stream-io/video-client';
 import {
   Restricted,
-  useCall,
-  useHasPermissions,
   useI18n,
   useLocalParticipant,
 } from '@stream-io/video-react-bindings';
@@ -12,6 +10,7 @@ import { useMediaDevices } from '../../core';
 import { DeviceSelectorAudioInput } from '../DeviceSettings';
 import { CompositeButton, IconButton } from '../Button';
 import { PermissionNotification } from '../Notification';
+import { useToggleAudioMuteState } from '../../hooks';
 
 export type ToggleAudioPreviewButtonProps = {
   caption?: string;
@@ -48,8 +47,6 @@ export type ToggleAudioPublishingButtonProps = {
 export const ToggleAudioPublishingButton = (
   props: ToggleAudioPublishingButtonProps,
 ) => {
-  const { publishAudioStream, stopPublishingAudio, setInitialAudioEnabled } =
-    useMediaDevices();
   const localParticipant = useLocalParticipant();
   const { t } = useI18n();
 
@@ -59,49 +56,8 @@ export const ToggleAudioPublishingButton = (
     SfuModels.TrackType.AUDIO,
   );
 
-  const call = useCall();
-  const hasPermission = useHasPermissions(OwnCapability.SEND_AUDIO);
-  const [isAwaitingApproval, setIsAwaitingApproval] = useState(false);
-  useEffect(() => {
-    if (hasPermission) {
-      setIsAwaitingApproval(false);
-    }
-  }, [hasPermission]);
-
-  const handleClick = useCallback(async () => {
-    if (
-      !hasPermission &&
-      call &&
-      call.permissionsContext.canRequest(OwnCapability.SEND_AUDIO)
-    ) {
-      setIsAwaitingApproval(true);
-      await call
-        .requestPermissions({
-          permissions: [OwnCapability.SEND_AUDIO],
-        })
-        .catch((reason) => {
-          console.log('RequestPermissions failed', reason);
-        });
-      return;
-    }
-    if (isAudioMute) {
-      if (hasPermission) {
-        setInitialAudioEnabled(true);
-        await publishAudioStream();
-      } else {
-        console.log('Cannot publish audio stream. Insufficient permissions.');
-      }
-    } else {
-      stopPublishingAudio();
-    }
-  }, [
-    call,
-    hasPermission,
-    isAudioMute,
-    publishAudioStream,
-    setInitialAudioEnabled,
-    stopPublishingAudio,
-  ]);
+  const { toggleAudioMuteState: handleClick, isAwaitingApproval } =
+    useToggleAudioMuteState();
 
   return (
     <Restricted requiredGrants={[OwnCapability.SEND_AUDIO]}>
