+++ packages/react-sdk/src/components/CallControls/ToggleVideoButton.tsx
@@ -1,17 +1,16 @@
-import { ComponentType, useCallback, useEffect, useState } from 'react';
+import { ComponentType } from 'react';
 import {
   Restricted,
-  useCall,
-  useHasPermissions,
   useI18n,
   useLocalParticipant,
 } from '@stream-io/video-react-bindings';
 
 import { OwnCapability, SfuModels } from '@stream-io/video-client';
 import { CompositeButton, IconButton } from '../Button/';
-import { DEVICE_STATE, useMediaDevices } from '../../core';
+import { useMediaDevices } from '../../core';
 import { DeviceSelectorVideo } from '../DeviceSettings';
 import { PermissionNotification } from '../Notification';
+import { useToggleVideoMuteState } from '../../hooks';
 
 export type ToggleVideoPreviewButtonProps = {
   caption?: string;
@@ -47,58 +46,17 @@ type ToggleVideoPublishingButtonProps = {
 export const ToggleVideoPublishingButton = (
   props: ToggleVideoPublishingButtonProps,
 ) => {
-  const { publishVideoStream, stopPublishingVideo, setInitialVideoState } =
-    useMediaDevices();
   const localParticipant = useLocalParticipant();
   const { t } = useI18n();
+
+  const { caption = t('Video'), Menu = DeviceSelectorVideo } = props;
+
   const isVideoMute = !localParticipant?.publishedTracks.includes(
     SfuModels.TrackType.VIDEO,
   );
-  const { caption = t('Video'), Menu = DeviceSelectorVideo } = props;
 
-  const call = useCall();
-  const hasPermission = useHasPermissions(OwnCapability.SEND_VIDEO);
-  const [isAwaitingApproval, setIsAwaitingApproval] = useState(false);
-  useEffect(() => {
-    if (hasPermission) {
-      setIsAwaitingApproval(false);
-    }
-  }, [hasPermission]);
-
-  const handleClick = useCallback(async () => {
-    if (
-      !hasPermission &&
-      call &&
-      call.permissionsContext.canRequest(OwnCapability.SEND_VIDEO)
-    ) {
-      setIsAwaitingApproval(true);
-      await call
-        .requestPermissions({
-          permissions: [OwnCapability.SEND_VIDEO],
-        })
-        .catch((reason) => {
-          console.log('RequestPermissions failed', reason);
-        });
-      return;
-    }
-    if (isVideoMute) {
-      if (hasPermission) {
-        setInitialVideoState(DEVICE_STATE.playing);
-        await publishVideoStream();
-      } else {
-        console.log('Cannot publish video. Insufficient permissions.');
-      }
-    } else {
-      stopPublishingVideo();
-    }
-  }, [
-    call,
-    hasPermission,
-    isVideoMute,
-    publishVideoStream,
-    setInitialVideoState,
-    stopPublishingVideo,
-  ]);
+  const { toggleVideoMuteState: handleClick, isAwaitingApproval } =
+    useToggleVideoMuteState();
 
   return (
     <Restricted requiredGrants={[OwnCapability.SEND_VIDEO]}>
