+++ packages/react-sdk/src/hooks/index.ts
@@ -1,2 +1,4 @@
 export * from './useFloatingUIPreset';
 export * from './useScrollPosition';
+export * from './useToggleAudioMuteState';
+export * from './useToggleVideoMuteState';
